+++ src/main/java/org/mbari/mxm/db/provider/ProviderApiType.java
@@ -1,6 +1,6 @@
 package org.mbari.mxm.db.provider;
 
 public enum ProviderApiType {
-  REST0,
+  REST,
   GRAPHQL,
 }