+++ src/main/java/org/mbari/mxm/provider_client/rest/MxmProviderClientRest.java
@@ -47,7 +47,7 @@ public class MxmProviderClientRest implements MxmProviderClient {
 
   @Override
   public ProviderApiType apiType() {
-    return ProviderApiType.REST0;
+    return ProviderApiType.REST;
   }
 
   @Override